Use floats instead of uint15/16
Performance seems to be about the same when operating on native buffers.
4x performance improvement when operating on RGBA float GeglBuffer.

+++ brushmodes.h
@@ -11,13 +11,13 @@ typedef struct {
     int y1; // Cordinates are pixel offsets into a tile
 } DabBounds;
 
-void draw_dab_pixels_BlendMode_Normal (uint16_t * mask,
-                                       uint16_t * rgba_buffer,
+void draw_dab_pixels_BlendMode_Normal (float * mask,
+                                       float * rgba_buffer,
                                        DabBounds * b,
-                                       uint16_t color_r,
-                                       uint16_t color_g,
-                                       uint16_t color_b,
-                                       uint16_t opacity);
+                                       float color_r,
+                                       float color_g,
+                                       float color_b,
+                                       float opacity);
 void
 draw_dab_pixels_BlendMode_Color (uint16_t * mask,
                                  uint16_t * rgba_buffer, // b=bottom, premult
@@ -26,23 +26,23 @@ draw_dab_pixels_BlendMode_Color (uint16_t * mask,
                                  uint16_t color_g,  // }-- a=top, !premult
                                  uint16_t color_b,  // }
                                  uint16_t opacity);
-void draw_dab_pixels_BlendMode_Normal_and_Eraser (uint16_t * mask,
-                                                  uint16_t * rgba,
+void draw_dab_pixels_BlendMode_Normal_and_Eraser (float * mask,
+                                                  float * rgba,
                                                   DabBounds *b,
-                                                  uint16_t color_r,
-                                                  uint16_t color_g,
-                                                  uint16_t color_b,
-                                                  uint16_t color_a,
-                                                  uint16_t opacity);
-void draw_dab_pixels_BlendMode_LockAlpha (uint16_t * mask,
-                                          uint16_t * rgba,
+                                                  float color_r,
+                                                  float color_g,
+                                                  float color_b,
+                                                  float color_a,
+                                                  float opacity);
+void draw_dab_pixels_BlendMode_LockAlpha (float *mask,
+                                          float *rgba,
                                           DabBounds *b,
-                                          uint16_t color_r,
-                                          uint16_t color_g,
-                                          uint16_t color_b,
-                                          uint16_t opacity);
-void get_color_pixels_accumulate (uint16_t * mask,
-                                  uint16_t * rgba,
+                                          float color_r,
+                                          float color_g,
+                                          float color_b,
+                                          float opacity);
+void get_color_pixels_accumulate (float * mask,
+                                  float * rgba,
                                   DabBounds *b,
                                   float * sum_weight,
                                   float * sum_r,
